[1.6.2] Fixed specs
