+++ spec/docscribe/cli/check_for_comments_spec.rb
@@ -6,10 +6,6 @@ require 'open3'
 require 'docscribe/cli'
 require 'docscribe/cli/check_for_comments'
 
-def resolve_config(raw, param_doc)
-  instance_double(Docscribe::Config, raw: raw, param_documentation: param_doc)
-end
-
 RSpec.describe Docscribe::CLI::CheckForComments do
   describe '.resolve_placeholders' do
     it 'returns default_message and param_documentation from config' do
