+++ spec/docscribe/types/yard/parser_spec.rb
@@ -4,14 +4,6 @@ require 'docscribe/types/yard/parser'
 require 'docscribe/types/yard/formatter'
 
 RSpec.describe Docscribe::Types::Yard do
-  def parse(string)
-    Docscribe::Types::Yard.parse(string)
-  end
-
-  def to_rbs(node)
-    Docscribe::Types::Yard::Formatter.to_rbs(node)
-  end
-
   describe '.parse' do
     it 'returns nil for nil' do
       expect(parse(nil)).to be_nil
