+++ spec/support/check_for_comments_helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module CheckForCommentsHelper
+  def resolve_config(raw, param_doc)
+    instance_double(Docscribe::Config, raw: raw, param_documentation: param_doc)
+  end
+end
