+++ spec/support/generate_helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module GenerateHelper
+  def run(*argv)
+    described_class.run(argv)
+  end
+end
