+++ spec/support/rbs_gen_helper.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module RbsGenHelper
+  def t(**kwargs)
+    described_class::YardTags.new(**kwargs)
+  end
+
+  def p(**kwargs)
+    described_class::ParamTag.new(**kwargs)
+  end
+
+  def d(**kwargs)
+    described_class::MethodDef.new(**kwargs)
+  end
+end
